After 3 hours of careful testing before merging...
...the very first real deploy failed, because of course it did.

So, somehow this wasn't how it was set up in the branch/PR (because deploy worked fine there), but then somehow magically it is now it was set up in main. All I've done is a simple merge, so I have *no* idea how it could be different, but I'm hungry and annoyed, so I give up.